Add lazy metadata for module metadata
This defers the loading/computing of some of the metadata information
for modules until it is accessed. This greatly improves performance
because most of the modules are not actually used and therefore their
metadata is also not used or needed. The time required to open and
process all the module files can be avoided.